requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
SQLAlchemy>=2.0.20
rq>=1.15.0
redis>=5.0.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
pyannote.audio>=3.1.0
speechbrain>=1.0.0
librosa>=0.10.1
soundfile>=0.12.1
torch>=2.0.0
torchaudio>=2.0.0
TTS>=0.22.0
langchain>=0.1.0
langchain-community>=0.0.34
langchain-chroma>=0.1.1
